requirements.txt
# Framework principal
Django==5.0.6

# Para criar a API RESTful
djangorestframework==3.15.1

# Para autenticação com JWT (JSON Web Tokens)
djangorestframework-simplejwt==5.3.1

# Para gerenciar permissões de acesso de outros domínios (CORS)
django-cors-headers==4.3.1

# Para carregar variáveis de ambiente (como a SECRET_KEY) do arquivo .env
python-dotenv==1.0.1

# Para gerar dados sintéticos (usado no seu comando generate_data)
Faker==25.8.0

# Driver para conectar o Django ao banco de dados MySQL
mysqlclient==2.2.4

# Biblioteca para manipulação de dados (leitura de CSV/Excel)
pandas==2.2.2
openpyxl==3.1.3